Introduced compression ratio parameter for Mixed Precision init (#133)
* Introduced compression ratio parameter for Mixed Precision init

It's used for choosing optimal mixed precision configuration for a given ratio.

Compression ratio of mixed precision quantization is calculated by relation to fully INT8 one.
Total compression for the model is sum of compression for each quantized layer, which is multiplication the layer's (Conv, Deconv, Linear) FLOPS and number of bits for its quantization. The ratio is used for estimation of performance boost for quantized model It's a better proxy for amount of calculation then number of parameters multiplied by bitwidth

* Added link to the full configuration file with template usage

* disclaimer about model specific params in template

* corrected articles, contractions, mixed precision-> mixed-precision