Removed custom SqueezeNet model for better user experience (#308)
* Correct configs for mixed precision models

After #300 *_hawq.json configs are propagation-based, but checkpoint are still for pattern-based quantization settings
That's why manual configs should be used to achieve a target accuracy

* Removed custom SqueezeNet model for better user experience

Originally we had a modified copy of SqueezeNet model to workaround a bug in ONNX exporter with converting MaxPool with ceil_mode=True.
This bug isn't actual now for torch 1.5 and there's almost identical SqueezeNet model in torchivision > 0.6.
That's why custom SqueezeNet was deleted as not needed to remove confusion.

There's no changes in the corresponding NNCF graph.
Previously trained checkpoints for custom SqueezeNet can be loaded and evaluated with SqueezeNet from torchvision. INT8 model has the same accuracy, mixed model is differ only by ~0.01 in maximum.