Dropped last data point in all DataLoaders to prevent issue with BN (#379)
There is a little chance that last data point may have a batch size equal to 1, which leads to an error:
```
ValueError: Expected more than 1 value per channel when training
```
We caught this error in sanity tests with CIFAR10. The dataset has 1000 data points.
There're 333 data points with batch_size=3 and the last one with batch_size=1. Training may fail in the end of epoch, which is not accepted for bigger datasets.