Remove pattern-based quantizer setup mode (#544)
* Remove pattern-based quantizer setup mode

* Return quantizable subgraph handling into insertion point graph creation

Does not help with GNMT quantization, however, since the current format
of specifying quantizable subgraph is not sufficient to describe the
blocks correctly, and the previous code state (with pattern-based)
applied the graph incorrectly - the matched pattern was not single
entry single exit.

* Remove get_weights_activation_pairs

The quantizer propagation setup mode supplies the required data
explicitly, no need to traverse the graph.

* Add option to specify input quantizer scale linking for propagation mode too