Fix for num_workers=0 for init_dataloader (#904)
Fix for LeGR slowdown in multiprocessing-distributed mode due to num_workers=0 for init loader (that was set to fix tests fails due to bug in Pytorch).
Currently, num workers for init loader in examples will be set as needed for distributed mode. And only in tests num_workers=0 will be used by redefining functions with data loaders creation.