Migrate to torch 1.9.1 (#953)
Fix for the failed tests in #947

All tests failed because of minor change in the MobileNet-v2 model: 
![image](https://user-images.githubusercontent.com/4014476/134982663-c19cf412-d9e8-4312-ad3f-1ae93c4c33e3.png)
`x.reshape(x.shape[0], -1)` is identical to `x=x.flatten(x,1)` according to the spec:

**Flatten**
Flattens input by reshaping it into a one-dimensional tensor. If start_dim or end_dim are passed, only dimensions starting with start_dim and ending with end_dim are flattened. The order of elements in input is unchanged.
https://pytorch.org/docs/stable/generated/torch.flatten.html

**Reshape**
Returns a tensor with the same data and number of elements as input, but with the specified shape. 
A single dimension may be -1, in which case it’s inferred from the remaining dimensions and the number of elements in input.
https://pytorch.org/docs/stable/generated/torch.reshape.html

The quantization setup is not affected and the same:
![image](https://user-images.githubusercontent.com/4014476/134983793-b1e5550f-a532-4acc-9536-412147e5b4b2.png)
![image](https://user-images.githubusercontent.com/4014476/134983949-2b599072-4a1f-461e-83b4-d5702d402a5e.png)

The fix in #953  - is just to regenerate reference graphs for mobilenet-v2 model.



Co-authored-by: Daniil Lyakhov <daniil.lyakhov@intel.com>