Mark and visualize integer edges in NNCF PT (#972)
### Changes

The integer edges are now being properly marked in NNCFGraph generated for the PyTorch backend, so that the solver can utilize this information. Furthermore the integer edges are now being displayed as dashed in `compressed_graph.dot` and `original_graph.dot` representations.

### Reason for changes

The models with integer tensors within their control flow graph will now be quantized properly - with less chance of a quantizer ending up quantizing an integer branch.

### Related tickets

N/A

### Tests

Added a test to build a graph on a toy model with integer tensors in the CFG, count the number of detected integer tensors and compare against the expected number.