Add NNCF repo commit hash to developer versions of the package (#1046)
### Changes

Now any version of the NNCF package set up in an editable mode will have "dev-editable" appended to its version; otherwise, if the package is built/installed via setup.py without passing a "--release" argument, the "dev-_hash of the current repo_" commit will be appended to the package version.

### Reason for changes

This allows for more accountability as to which NNCF version is used where.

### Related tickets

74164

### Tests

test_install (to be run)