Make dev versions conform to PEP 440 (#1050)
### Changes

The '-' in dev versions containing commit hash was changed to '+'.

### Reason for changes

Previous version format was not conforming to PEP 440, and the `pip install git+` path of installation failed to install from a wheel. This was not fatal, but a nasty warning was being spawned and a deprecated installation path had to be involved.

### Related tickets

N/A

### Tests

Extended test_install with the `pip install git+` scenarios.