Cast scale to the input_low/input_range dtype (#1088)
### Changes

In the asymmetric quantizer execution, add a safety cast of the intermediate calculations to the precision of the tensors that are input into the quantizer.

### Reason for changes

Due to unknown reason the intermediate calculations with FP16 torch values result in FP32 outputs, which breaks downstream code that expects that all values involved have one and the same dtype.

### Related tickets

70246

### Tests

Cannot be reproduced in laboratory conditions even if the same tensor values are used.