Changed normalization parameters for CIFAR-10 images (#1091)
### Changes

Changed mean/std values for pre-processing input images from CIFAR10 dataset

### Reason for changes

It helps to reuse models pretrained on CIFAR10 from:
https://github.com/huyvnphan/PyTorch_CIFAR10/blob/master/data.py#L16
This models are used for e2e validation of BootstrapNAS feature #1079 

### Related tickets

### Tests

It affects backward compatibility tests that compare actual test accuracy with 'best_acc1' value in the checkpoint.
The `best_acc1` value was overridden to correspond to the introduced pre-processing.