Corrected typo to fix broken TF pre-commit (#1102)
### Changes

Corrected typo in a TF test to fix degradation in the TF precommit with new version of pytest.
Seems like previous runs was with old version of pytest (6.2.5) and it ignored this typo.
Currently, CI uses newer version - 7.0.0 and it's failing.

### Reason for changes

to fix degradation in the TF precommit

### Related tickets

### Tests

tf precommit should be green