Correct statistics (filter pruning and sparsity) (#1098)
Changes

Statistics of the pruned model are now collected from the model (using get_nncf_operations) but not from the graph (affected when loading from a checkpoint, the graph is not getting updated, but the model is).

When loading from a checkpoint the current_level of Scheduler is recalculated.

current_step and current_epoch attributes of BaseCompressionScheduler become private (just getters are available - not possible to update from outside). Tests are updated in accordance (no direct set of current_step and current_epoch).
Reason for changes

Printed statistics of the model and algorithm are not correct while loading from the checkpoint.
Tests

Updated.