Introduce ONNXTargetPoint, StatisticPoint. (#1241)
### Changes

1) Introduce a class for storing the necessary statistics information - StatisticPoint.
2) Introduce ONNXTargetPoint to unify with the NNCF common part - TargetPoint. 

3) ONNXMinMaxQuantization now treats quantizers like ONNXTargetPoint, which has TargetType field demonstrates where the quantizer should be placed (before, after operation or quantize weights).

4) Added new method for Engine - 
```
@abstractmethod
def _register_statistics(self, outputs: NNCFData, statistic_points: StatisticPointsContainer) -> None:
    """
    Does mapping from the provided output and statistics_points to register statistics.
    ""
```
This method is essential to make a mapping between models output and the statstic_points. 
Reviewers, please take a look at this.

5) Update ONNXModelTransformer - now it operates with ONNXTargetPoint, add new ONNXInsertionCommand - ONNXOutputInsertionCommand

6) Update StatisticsAggregator: unify collect_statistics and add some backend-specific abstractmethod.

### Reason for changes

This PR makes preparation steps for implementing FastBiasCorrection algorithm.

### Related tickets

76189

### Tests

The order of quantizers in model graphs were updated according to the new logic that weight and activation quantizes are stored in one place. So the order in graphs were updated.

The test with insertion extra outputs to the model is added