Introduced 2 strategies to remove overlapping building blocks (#1281)
Resolved numerous issues with automatic search of building blocks to enable pruning of transformers and to not break the related NAS functionality.

### Changes

Originally, this PR fixes #1153 issue by introducing 2 new strategies for removing overlapped building blocks.
The first strategy is `KEEP_SMALL` that gives a preference to small blocks. It starts from the smallest block and filters all blocks that intersect or include it. Then it finds the next smallest block from the remaining ones and repeats the procedure.
This strategy can consider the blocks of the given type first, if the list of types is provided. That's helpful to find all MHSA and FF blocks in typical transformers: BERT, ViT and Wave2Vec.

The default strategy is `KEEP_SEQUENTIAL` that gives a preference to sequential blocks, which follow each other in the model. This strategy should be helpful for Progressive Shrinking Algorithm in BootstrapNAS.

In order to pass CI with 2 new strategies, it also fixes invalid logic with starting node in the automatic search of building blocks. 
The issue is described in https://github.com/openvinotoolkit/nncf/pull/1270

### Reason for changes

The functionality is needed for pruning transformers and resolve errors on inference of SuperNet for some models: ShuffleNet, EfficientNet and MobileNet-v3.

### Related tickets

#1153, 85935, 60403, 80679, 77168, 

### Tests
new tests for overlapping
new tests for transformers
all previous NAS and search building block tests