Corrected extraction of layer attributes in case of weight_norm (#1309)
### Changes

Extract `requires_grad` attribute for weight_g and weight_v if there exist.

### Reason for changes

[torch.nn.utils.weight_norm](https://pytorch.org/docs/stable/_modules/torch/nn/utils/weight_norm.html#weight_norm)
“patches” ordinary module by removing weight and adding weight_g and weight_v.
The patched module looks as ordinary one. There's even a weight attribute, but it's just a tensor, not a learnable parameter anymore.
For conv modules, we always check `requires_grad` of weight attribute, and it's false after applying `weight_norm` utility. 
We should consider 'weight_g' or 'weight_v' if there exist.
The patched module looks as a frozen one, because weight is a tensor and requires_grad=False, but actually it's not because there's a modified functionality via weight_g and weight_v that are learnable.

### Related tickets

closes #1308

### Tests

random_multi_elasticity for TCN model