Make OrderedEnum int based (#1374)
### Changes

Removed `OrderedEnum` and replaced its usages by `IntEnum`.

### Reason for changes

In general `OrderedEnum` should base on `IntEnum` because a method \_\_lt\_\_ is implemented for it where enum values are compared with "<" sign. This is not critical, but currently it complicates development of #1343. There, when an object of type `CompressionStage(OrderedEnum)` is being serialized by json, it throws an error, because it is ambiguous how to serialize a general enum object. Making `OrderedEnum` be based from `IntEnum` resolves this issue.

**Edit:** decided to remove `OrderedEnum` altogether and use `IntEnum` instead.