Rework NNCF module placement to track all access scopes of the replaced module (#1077)
### Changes

When replacing NNCF modules, keep track of all scopes through which the module is accessible in the original model object. When building the DynamicGraph, keep track of the module `id` that the operation is executed in, and reuse this information to detect shared-layer nodes in NNCFGraph. If there are multiple scopes for the module that contains a given operation (represented by the node in the graph), then the node is associated with the lexicographically first scope of the containing module.

### Reason for changes

Some PT models may be written in such a fashion that a given submodule calls a certain other submodule in the model, but at the same time this first submodule does not contain the second in the storage hierarchy (see the test case in the PR for an explicit example of such a model). In this case the storage scope does not coincide with the execution scope of the operations; this also causes problems during module traversal in `replace_modules`.

### Related tickets

75373

### Tests

tests.torch.test_nncf_network.test_wrapping_of_indirect_module_operations