Update eval results in tables across READMEs and align TF model names with PT (#1507)
### Changes
Updated metric values in tables and in reference files. Made TF models
conform to the way of naming in PT. Enabled in-place editing for the
`tools/update_eval_results.py` script to do this automatically using
only the `results.json` file from CI.

### Reason for changes
Catching up to the metric drift.

### Related tickets
N/A

### Tests
E2E TF for the model renaming part.