[Pruning] Fixed statistics computation for pruned layers (#1597)
### Summary
For PT filter pruning, before `compression_ctrl.export_model()` call the pruned layer statistics are printed. They however are currently empty and do not reflect the actual pruned state of the model.

### Reason
This is due the fact that `pruning_level_for_filters()` method used to retrieve statistics returns incorrect values. This in turn is due the fact that it relies on the number of zeros in the model weights. But weights are zeroed (pruned) only during model export and actual model weights are not zeroed during runtime.

### What's changed
Removed `pruning_level_for_weight()` and `pruning_level_for_filters()` altogether. Now `pruning_level_for_mask()` is the only method to compute the statistic.


#### Before
![image](https://user-images.githubusercontent.com/23343961/220954948-8498ea47-cf15-4801-a014-7852b08914bd.png)
#### After
![image](https://user-images.githubusercontent.com/23343961/220955021-72a98967-fa8b-4f4d-a5f0-d2a1d09afb88.png)