Added new AA config on CIFAR10 and ResNet18 (#1641)
### Changes

- Added new Accuracy Aware config for quantization + pruning scenario on
ResNet18 model and CIFAR10 dataset
- Added logic for downloading models

Results are given below. The AA metrics were averaged over 10 runs.
|   | Accuracy  | Compression rate  |  Duration |  
|---|---|---|---|
| Baseline FP32 | 88.61%  | 0%  | -  | 
| AA INT8 + Pruning|  87.78% | 58.13%  | 44.3 min  |


### Reason for changes

Existing AA configs run on ImageNet and take multiple days to finish
which makes it hard for new users to get familiar with the approach and
use it. The new config runs 45 mins on average with 8 workers on
i9-10980XE CPU and single RTX 3090 GPU.