Rework NNCFNetwork for better integration (#1584)
### Changes

`NNCFNetwork` was refactored from a wrapper-approach to a mixin-like approach.

### Reason for changes
`NNCFNetwork` is the object we return to the user as a compression-enabled counterpart to the original model object, and we want this object to look and feel the same in terms of further usage of this object in the training pipeline. In the wrapper approach, the object no longer had the original class, attribute accesses had to be forwarded to the wrapped object, and there was a possibility of name collision. With the mixin-approach, the returned model object has a dynamically extended class, which, however, looks the same as the original class in terms of `isinstance` checks, and also hashing/comparisons, and even has the same `forward` signature, which will be useful since we are switching to the `prepare_for_inference` approach where we don't control the export process.

### Related tickets

79448

### Tests

tests/torch/test_nncf_network.py