Make access to original_unbound_forward through original class reference (#1802)
### Changes

Made so that we access the unbound forward from a reference to the original model class instead of storing the reference to the unbound forward itself because the original class forward may be overridden by some 3rd party logic.

For example, during export of mm-based models to ONNX using mmdeploy library, the original forward method of the model is temporarily replaced during export. Moreover, in such case the forward signature needs to be hidden by a user beforehand by wrapping it with a function with `(*args, **kwargs)` as its arguments.

### Reason for changes

Enabling export of models compressed with NNCF by mmdeploy.

### Tests

Added two tests covering new functionality.