Force range initialization to be applied in eval mode (#1828)
### Changes

Force range initialization to be applied in eval mode.

### Reason for changes

If model in train mode is given to `create_compressed_model()`, the
graph is built for the model in eval mode, but range initialization
statistics are collected in train mode. This may result in wrong
statistics collection if model behaves differently in train and eval
modes.

### Tests

Added a test failing before the fix.