Ignore 0-th port of baddbmm in quantization (#1852)
Add a field to `OperatorMetatype` that specifies input ports to be
ignored and use it to disable quantization on the 0-th input port of the
`torch.baddbmm` operation.

The 0-th input for baddbmm is a bias addition in essence, and if the
input tensor is a constant, then adding a quantizer there breaks down a
hardware-fused pattern and reduces performance. It is still possible
that the 0-th input may be a runtime-dynamic activation tensor, in which
case the quantization of this input would probably still be relevant,
but right now there is no way in NNCF PT to discern activation and
constant inputs.

110848

test_synthetic_model_quantization