[OV] Update tests references (#1885)
### Changes

- **Reverted**. Added `pytest.xfail` for `mobilenet-v3-small-1.0-224-tf`
model due to OMZ failures (fixed in master -
https://github.com/openvinotoolkit/open_model_zoo/pull/3777);
- Updated `googlenet-v3`, `mobilenet-v2`, `resnet-18` architecture
examples (.dot) due to input **processing pattern changes** (see below);
- Updated `yolo-v4-tiny` and `mobilenet-v3-small` architecture examples
(.dot) and scales (.json) due to layers names changes that provides new
MO TF frontend (see below);
- Updated `mobilenet-v2`, `resnet-18` architecture examples scales
(.json) due to input **processing pattern changes** (see below);
- Updated required OpenVINO package version to 2023.0.0;
- Updated OMZ package version;

### Reason for changes

- Red CI

### Details
- **OV 2023.0.0** output graph (resnet-18):

![image](https://github.com/openvinotoolkit/nncf/assets/32988081/a76a3f0c-c83a-45d9-870d-ed4859bb26bc)
- **OV 2022.3.0** output graph (resnet-18):

![image](https://github.com/openvinotoolkit/nncf/assets/32988081/f550dd56-516a-4ccd-b692-a219e0a394b6)

As we can see, **only** the output processing subgraph was changed. 
For example, the FakeQuantize node for the second Convolution does not
change its parameters.

In addition to the fact that the names have changed due to the new front
end, the position of the layers has also changed.
For example:
- **OV 2023.0.0** original graph (mobilenet-v3-small):

![image](https://github.com/openvinotoolkit/nncf/assets/32988081/543b3a04-d5ff-4962-8dae-4ac7ca36a688)
- **OV 2022.3.0** original graph (mobilenet-v3-small):

![image](https://github.com/openvinotoolkit/nncf/assets/32988081/30b6285d-4cf6-44e8-97e4-5c9489af26ff)

This is the reason for changing the number of FakeQuantize layers (+1
activation for separate Add) in the model (22.3 vs. 23.0).