Corrected condition for matching recurrent nodes (#1998)
### Changes

The node is considered within iteration scope if the correspondent module name fully matches the registered names of registered recurrent modules (LSTM, GRU cells). 
Previously, there was a less strict rule: the name of the iteration module should include the name of the considered module.

### Reason for changes

The problem appeared with `RecurrentDecoder` with a customer model: 
https://github.com/PeterL1n/RobustVideoMatting/blob/master/model/model.py#L26C28-L26C44
FQ should be propagated up through `concat` and `strided slice`, but it was mistakenly considered within iteration scope and added to ignored scope for quantization.
The concat's scope is `205 MattingNetwork/RecurrentDecoder[decoder2]/OutputBlock[decode0]/cat_0`
It matched with one of the registered iteration scopes – `Recurrent`. With a corrected condition, it's not matched.
![image](https://github.com/openvinotoolkit/nncf/assets/4014476/84d1a79a-45ad-4713-8064-20fa9f0fa9fa)
![image](https://github.com/openvinotoolkit/nncf/assets/4014476/b0108a68-ed3f-4554-b55a-eed8a4a95259)


### Related tickets

112934

### Tests

synthetic tests for model with Recurrent in the name