Remove pylint (#2205)
### Changes
Pylint was removed from the make targets and from code. Added `.pylintrc` to `.gitignore` so that developers could have their own `.pylintrc` if they still want to use pylint in informational capacity locally.

### Reason for changes
Following through with the decision to use ruff as the main linter. Most of the pylint rules not covered by ruff will be covered by mypy in the future.

### Related tickets
N/A

### Tests
Existing test scope