Removed mlflow usage (#2395)
### Changes

removed usage of mlflow

### Reason for changes

AFAIK, it's not used during the development in the team and there was no
interest from customers to extend it.
Also, it's extra dependency and often leads to errors in validation,
that required several workarounds (SafeMLFLow and `mock_mlflow`):
```
E           Exception: Run with UUID 14ae21d2c1d349839bca0371583cfc78 is already active. To start a new run, first end the current run with mlflow.end_run(). To start a nested run, call start_run with nested=True
```
### Related tickets

<!--- Post the numerical ID of the ticket, if available -->

### Tests

- [x] nightly torch: WIP build 108
- [x] nightly torch: WIP build 110