Hide an irrelevant warning during tune_range call (#2415)
When `tune_range()` is called with a `left_border` being equal to 0.
This leads to the following warning
```
/home/nsavel/workspace/nncf/nncf/experimental/tensor/tensor.py:84: RuntimeWarning: invalid value encountered in multiply
  return Tensor(self.data * unwrap_tensor_data(other))
```

at this line:

```
rb = fns.where(qval > 0.0, (qval - level_high) / qval * left_border, right_border)
```

When we divide by `qval` which equals 0, this results in a NaN during
computation. At the same time if `qval` is zero then `where` condition
does not hold and `rb==right_border`. Seems like numpy computes both
results nevertheless of the condition.

### Changes

Do not show the unnecessary warning because it is confusing.

### Reason for changes

Improving UX

### Tests

Added a test for `tune_range`. Not sure about the location, feel free to
suggest a better one.

---------

Co-authored-by: Alexander Suslov <alexander.suslov@intel.com>