Port `.take` functionality for Good First Issues from OpenVINO repository (#2544)
### Changes / Reason for changes

Merging this workflow allows open source contributors to comment .take
under a Good First Issue and automatically pick it up. The community is
already accustomed to that and currently it does not work in NNCF GFIs.

### Related tickets

N/A

### Tests

`.take` has been proven to be a convenient feature for both project
maintainers and open source contributors. Here it is in OpenVINO repo:
https://github.com/openvinotoolkit/openvino/blob/master/.github/workflows/assign_issue.yml.