Fixed netron visualization (#2567)
### Changes

This PR solves #2552 by fixing the Netron visualization. The function
`save_for_netron` now produces an XML file that can be correctly opened
by Netron. To achieve this, a dummy `dtype` conversion has been
introduced, as discussed in #2552. This conversion maps the nncf dtype
`Dtype.FLOAT` to `f32` and `DType.INTEGER` to `i32`.

The `precision` parameter of the class `PortDesc` now is no longer
optional as it's always available and it's necessary to produce a
working XML file.

In addition, I added a docstring for all the functions/classes and
implemented tests for the following methods:

- `get_graph_desc()`
- `PortDesc.as_xml_element()`
- `NodeDesc.as_xml_element()`
- `EdgeDesc.as_xml_element()`
 

### Reason for changes

<!--- Why should the change be applied -->
It was not possible to open XML files produced by the function
`save_for_netron` due to the error: `Error loading OpenVINO model.
Unsupported precision 'undefined'`

### Related tickets

N/A

### Tests

To validate the accuracy of the modifications, I created Netron XML
files from multiple ONNX models. The visualization of these models in
Netron was successful, confirming the effectiveness of the changes.