Tests for torch device and dtype utils (#2580)
### Changes

This PR addresses #2579.
- Add tests for torch device utils. The tests consider the case in which
the model has no parameters, has all parameters on CPU, has all
parameters on CUDA, and has parameters placed on different devices. In
the latter, the parameters are moved on different devices randomly.
- Add tests for torch `dtype` utils. The case in which the model has no
parameters is also considered.
- Created a test torch helper class `EmptyModel` to take into account
the case in which the model has no parameters at all.
- Add docstrings in utils.

### Tests

I compared all the results manually checking for their correctness. The
code is also compliant with the coding style having been verified with
`pre-commit run`

---------

Co-authored-by: Daniil Lyakhov <daniil.lyakhov@intel.com>