Add early stopping for memory monitoring (#2897)
### Changes

Stop addition of new memory values as soon as possible in order to
correctly represent the tracked task.

### Reason for changes

When running memory monitor as a CLI tool, sometimes it may happen that
the last value is recorded in the main process, i.e. when the child
process has already died. This is represented as a drop in memory,
because the child process memory is deallocated. This PR tries to avoid
such behavior.