Add weighted progress tracking for weight compression (#2892)
### Changes

During nncf weight compression, `rich` progress bar is used to display
the progress. In this PR, progress bar is changed to be weighted
according to model weights. With these changes, each weight contributes
proportional amount of percent to the progress bar.

Iteration number was removed from weight compression progress bar to
avoid confusion between different speeds in percent and iteration
coordinates. For example now a single weight might contribute 5-10% to
the whole progress.

### Reason for changes

The time it takes to compress a weight is roughly proportional to its
size, so incrementing the progress by 1 for each weight is not ideal.

Especially after #2803 when weight sorting was added. Now, the largest
weights come first and the smallest ones are at the end of the
compression. This leads to misleading time estimation when progress
contribution from every weight is equal.

Weights sizes for tinyllama-1.1b for reference:

![weight_size_hist](https://github.com/user-attachments/assets/30ba1e1b-0fc5-4d6b-84db-948362672bf2)


![weight_size_cumsum_hist](https://github.com/user-attachments/assets/b00e79e8-5000-44a4-97a5-4102c9aed0ae)