Preserve the constant subgraph (#2902)
### Reason for changes

Removing the ShapeOf subgraph may lead to the removal of some input
edges of operations, and QPS doesn't place FQ on these edges. In this
case, the operations are quantized incorrectly because not all required
inputs are quantized. These operations don't change their runtime
precision in the execution graph, which means we do not have a fully
quantized model.

### Changes

To resolve the described problem, this PR introduces the following
algorithm:

- **Step 1:** Find all nodes belonging to the ShapeOf subgraphs in the
graph. Let's denote the found nodes as $S$. At this step, we don't
remove the found subgraphs; we only store the nodes they consist of. See
`find_shapeof_subgraphs()` method.

- **Step 2:** Check the producers for the specified nodes (currently, we
are considering only nodes of the LSTMSequence and Convolution types).
If any of these producers are in set $S$, it means that we will lose
this input edge when the ShapeOf subgraph is removed. Therefore, we try
to exclude float subgraphs from this ShapeOf subgraph. Let's denote the
nodes that are part of such float subgraphs as $P$. We can see that $P
\subset S$. See `find_preserved_nodes()` method.

- **Step 3:** Find all nodes belonging to the Constant subgraphs in the
graph. Let's denote the found nodes as $C$. At this step, we don't
remove the found subgraphs; we only store the nodes they consist of. See
`find_constant_subgraphs()` method.

- **Step 4:** Remove $S \cup C - P$ nodes from the graph.

### Related tickets

- 146088