Fix torch compile applied to model forward (#2932)
### Changes

Check whether model is compiled based on `"_torchdynamo_orig_callable"`
property of the model forward.

### Reason for changes

`torch.compile` can be applied not only to the model itself, but to
`forward()` method only. For example:
```
model.forward = torch.compile(model.forward)
```
In this case the model itself doesn't change and it won't be an instance
of `torch._dynamo.OptimizedModule`.

### Related tickets

143796

### Tests

Added test when `torch.compile` is applied this way. It does not fail
without the fix though, because the issue is sporadic.


### Relates to
#2665, #2719